(Refactor) Move mdbook.rs to mod.rs
